message/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
pymongo==4.1.1
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
urllib3==1.26.9
